+++ train_flux/config.yaml
@@ -1,6 +1,6 @@
 model_path: "black-forest-labs/FLUX.1-dev"
 dtype: "bfloat16"
-cache_dir: "/fsx/sayak/.cache"
+cache_dir: "CACHE_DIR"
 
 model:
   union_cond_attn: true
@@ -10,7 +10,7 @@ model:
 train:
   batch_size: 8
   accumulate_grad_batches: 1
-  dataloader_workers: 2
+  dataloader_workers: 8
   save_interval: 2000
   sample_interval: 2000
   max_steps: -1
@@ -23,7 +23,7 @@ train:
   resume_training_from_checkpoint_path: ""
   dataset:
     type: "img"
-    path: "pipe:curl -s -f -L https://huggingface.co/datasets/diffusion-cot/GenRef-wds/resolve/main/genref_{0..16}.tar"
+    path: "pipe:curl -s -f -L https://huggingface.co/datasets/diffusion-cot/GenRef-wds/resolve/main/genref_{0..208}.tar"
     split_ratios: {
       "general": [0.1, 0.3],
       "length": [0.1, 0.3],
@@ -32,15 +32,15 @@ train:
     }
     training_stages: [0, 5000]
     root_dir: ""
-    val_path: {
-      "general": ["/mnt/petrelfs/zhuole/ReflectionFlow_/val.json"]
-    }
-    val_root_dir: ""
-    condition_size: 512
-    target_size: 1024
-    drop_text_prob: 0.1
-    drop_image_prob: 0.1
-    drop_reflection_prob: 0.1
+    # val_path: {
+    #   "general": "VAL_TARS"
+    # }
+    # val_root_dir: ""
+    # condition_size: 512
+    # target_size: 1024
+    # drop_text_prob: 0.1
+    # drop_image_prob: 0.1
+    # drop_reflection_prob: 0.1
 
   wandb:
     project: "ReflectionFlow"
